Add plain text words to dictionary, not original
For example, when in LaTeX "Test\-Word" was underlined because
"Test-Word" could not be recognized, "Test\-Word" would have been
added to the dictionary. This commit changes this to be
"Test-Word" by using the "plain text" (text given to LanguageTool)
instead of the "original text" (code in editor) in LanguageTool
terms.